POI
Fetching POIs from Foursquare and adding them to the database. If a the POIs in a given radius, calculated according to the haversine formula, are already in the database then they will be retrieved from there, saving API calls to Foursquare and time for the user.

According to Foursquare TOS, the data must be cached for only a day. A purge task has been asigned to remove any data older than a day.